Remove xgboost4j-gpu, move them into xgboost4j-spark-gpu
