+++ jvm-packages/pom.xml
@@ -97,9 +97,8 @@
       </properties>
       <modules>
         <module>xgboost4j</module>
-        <module>xgboost4j-gpu</module>
-<!--        <module>xgboost4j-spark</module>-->
-<!--        <module>xgboost4j-spark-gpu</module>-->
+        <module>xgboost4j-spark</module>
+        <module>xgboost4j-spark-gpu</module>
       </modules>
     </profile>
