+++ jvm-packages/xgboost4j-spark-gpu/src/main/resources/META-INF/services/ml.dmlc.xgboost4j.scala.spark.PreXGBoostProvider
@@ -1,1 +1,1 @@
-ml.dmlc.xgboost4j.scala.rapids.spark.GpuPreXGBoost
+ml.dmlc.xgboost4j.scala.GpuPreXGBoost
