--- jvm-packages/xgboost4j-spark-gpu/src/main/scala/org
@@ -1,1 +1,0 @@
-../../../../xgboost4j-spark/src/main/scala/org
