--- jvm-packages/xgboost4j-spark-gpu/src/test/resources
@@ -1,1 +1,0 @@
-../../../xgboost4j-spark/src/test/resources
